Create core circuit tests
